+++ test/fixtures/app/xyz/foo.go
@@ -1,0 +1,9 @@
+package xyz
+
+import (
+	"golang.org/x/crypto/ssh"
+)
+
+func Foo() {
+	_ = ssh.InsecureIgnoreHostKey()
+}