+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Go AST Scanner analyzer changelog
 
+## 10-8-stable
+- Rewrite using Go and analyzers common library
+
 ## v0.1.0
 - initial release